Move TestCase migration tool to java/de/jflex/migration/testcase/
Subdirectory in preparation of #859
