+++ java/de/jflex/migration/testcase/MigrationException.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019-2020 Google, LLC.
+ * Copyright (C) 2019-2021 Google, LLC.
  *
  * License: https://opensource.org/licenses/BSD-3-Clause
  *
@@ -23,7 +23,7 @@
  * WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY
  * WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package de.jflex.migration;
+package de.jflex.migration.testcase;
 
 /**
  * Thrown when a severe problem happened while migrating a maven test case to bazel using {@code
